remove unused pages, make new console/voting pages the only/default pages
