+++ src/pages/MemberConsole.js
@@ -1,89 +1,101 @@
-import { useState, useEffect, useCallback } from 'react';
-import { div, h, img } from 'react-hyperscript-helpers';
-import { DAR } from '../libs/ajax';
+import {useState, useEffect, useRef, useCallback } from 'react';
 import SearchBar from '../components/SearchBar';
-import { Notifications, getElectionDate, processElectionStatus, darSearchHandler } from '../libs/utils';
+import { Collections, User } from '../libs/ajax';
+import { Notifications, searchOnFilteredList, getSearchFilterFunctions } from '../libs/utils';
 import { Styles } from '../libs/theme';
-import DarTable from '../components/dar_table/DarTable';
+import {h, div, img} from 'react-hyperscript-helpers';
 import lockIcon from '../images/lock-icon.png';
-import { updateLists as updateListsInit } from '../libs/utils';
-import { tableHeaderTemplate, tableRowLoadingTemplate } from '../components/dar_table/DarTable';
-import DarTableSkeletonLoader from '../components/TableSkeletonLoader';
-import { isNil } from 'lodash/fp';
+import { DarCollectionTable, DarCollectionTableColumnOptions } from '../components/dar_collection_table/DarCollectionTable';
 import { consoleTypes } from '../components/dar_table/DarTableActions';
 
 export default function NewMemberConsole(props) {
-  const [electionList, setElectionList] = useState([]);
+  const [collections, setCollections] = useState([]);
   const [filteredList, setFilteredList] = useState([]);
-  const [tableSize, setTableSize] = useState(10);
-  const [descendantOrder, setDescendantOrder] = useState(false);
-  const [currentPage, setCurrentPage] = useState(1);
-  const [isLoading, setIsLoading] = useState(false);
+  const [relevantDatasets, setRelevantDatasets] = useState();
+  const [isLoading, setIsLoading] = useState(true);
+  const searchRef = useRef('');
+  const filterFn = getSearchFilterFunctions().darCollections;
+  const { history } = props;
+
+  const handleSearchChange = useCallback(
+    (searchTerms) =>
+      searchOnFilteredList(searchTerms, collections, filterFn, setFilteredList),
+    [collections, filterFn]
+  );
 
   useEffect(() => {
     const init = async () => {
       try {
-        setIsLoading(true);
-        const pendingList = await DAR.getDataAccessManageV2();
-        const openElectionList = (pendingList).filter((i) => {
-          return isNil(i.election) ? false : i.election.status === 'Open';
-        });
-        setElectionList(openElectionList);
-        setFilteredList(openElectionList);
+        const [collections, datasets] = await Promise.all([
+          Collections.getCollectionsByRoleName('member'),
+          User.getUserRelevantDatasets(), //still need this on this console for status cell
+        ]);
+        setCollections(collections);
+        setRelevantDatasets(datasets);
+        setFilteredList(collections);
         setIsLoading(false);
       } catch (error) {
-        Notifications.showError({text: 'Error: Unable to retrieve data requests from server'});
+        Notifications.showError({
+          text: 'Error initializing Collections table',
+        });
       }
     };
     init();
   }, []);
 
-  const getUpdateLists = useCallback(() => {
-    return updateListsInit(filteredList, setFilteredList, electionList, setElectionList, currentPage, tableSize);
-  }, [filteredList, electionList, currentPage, tableSize]);
+  const goToVote = useCallback((collectionId) => history.push(`/dar_collection/${collectionId}`), [history]);
 
-
-  const handleSearchChange = darSearchHandler(electionList, setFilteredList, setCurrentPage);
-
-  return (
-    div({style: Styles.PAGE}, [
-      div({style: {display: 'flex', justifyContent: 'space-between'}}, [
-        div({className: 'left-header-section', style: Styles.LEFT_HEADER_SECTION}, [
-          div({style: Styles.ICON_CONTAINER}, [
+  return div({ style: Styles.PAGE }, [
+    div({ style: { display: 'flex', justifyContent: 'space-between', width: '112%', marginLeft: '-6%', padding: '0 2.5%' } }, [
+      div(
+        { className: 'left-header-section', style: Styles.LEFT_HEADER_SECTION },
+        [
+          div({ style: Styles.ICON_CONTAINER }, [
             img({
               id: 'lock-icon',
               src: lockIcon,
-              style: Styles.HEADER_IMG
-            })
+              style: Styles.HEADER_IMG,
+            }),
           ]),
-          div({style: Styles.HEADER_CONTAINER}, [
-            div({style: Styles.TITLE}, ['DAC Member Console']),
-            div({style: Object.assign({}, Styles.MEDIUM_DESCRIPTION, {fontSize: '18px'})}, ['Select and manage Data Access Requests for DAC review'])
+          div({ style: Styles.HEADER_CONTAINER }, [
+            div({ style: {
+              fontFamily: 'Montserrat',
+              fontWeight: 600,
+              fontSize: '2.8rem'
+            } }, [
+              `My DAC's Data Access Requests`,
+            ]),
+            div(
+              {
+                style: {
+                  fontFamily: 'Montserrat',
+                  fontSize: '1.6rem'
+                },
+              },
+              ['Vote on Data Access Request for DAC Review']
+            ),
           ]),
-        ]),
-        h(SearchBar, {handleSearchChange, currentPage})
-      ]),
-      h(DarTable, {
-        getUpdateLists,
-        filteredList,
-        setFilteredList,
-        descendantOrder,
-        setDescendantOrder,
-        history: props.history,
-        processElectionStatus,
-        getElectionDate,
-        consoleType: consoleTypes.MEMBER,
-        currentPage,
-        setCurrentPage,
-        tableSize,
-        setTableSize,
-        isRendered: !isLoading
-      }),
-      h(DarTableSkeletonLoader, {
-        isRendered: isLoading,
-        tableHeaderTemplate: tableHeaderTemplate(consoleTypes.MEMBER),
-        tableRowLoadingTemplate: tableRowLoadingTemplate(consoleTypes.MEMBER)
-      })
-    ])
-  );
+        ]
+      ),
+      h(SearchBar, { handleSearchChange, searchRef }),
+    ]),
+    h(DarCollectionTable, {
+      collections: filteredList,
+      columns: [
+        DarCollectionTableColumnOptions.DAR_CODE,
+        DarCollectionTableColumnOptions.NAME,
+        DarCollectionTableColumnOptions.SUBMISSION_DATE,
+        DarCollectionTableColumnOptions.RESEARCHER,
+        DarCollectionTableColumnOptions.INSTITUTION,
+        DarCollectionTableColumnOptions.DATASET_COUNT,
+        DarCollectionTableColumnOptions.STATUS,
+        DarCollectionTableColumnOptions.ACTIONS,
+      ],
+      isLoading,
+      relevantDatasets,
+      reviseCollection: null,
+      goToVote,
+      consoleType: consoleTypes.MEMBER
+    }),
+  ]);
 }
